% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.key.R
\name{as.key}
\alias{as.key}
\title{Convert data.frame to a Hash Key}
\usage{
as.key(x, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} with the first column containing
polarized words and the second containing polarity values.}

\item{\ldots}{ignored.}
}
\value{
Returns a \pkg{data.table} object that can be used as a hash key.
}
\description{
Create your own hash keys from a data frame for use in key arguments such as
\code{polarity_dt} in the \code{sentiment} function.
}
\examples{
key <- data.frame(
    words = sample(LETTERS),
    polarity = rnorm(26),
    stringsAsFactors = FALSE
)

(mykey <- as.key(key))

# Looking up values
mykey[c("A", "K")][[2]]

# Using syuzhet's sentiment lexicons
\dontrun{
library(syuzhet)
as.key(syuzhet:::bing)
as.key(syuzhet:::afinn)
nrc <- data.frame(
    words = rownames(syuzhet:::nrc),
    polarity = syuzhet:::nrc[, "positive"] - syuzhet:::nrc[, "negative"],
    stringsAsFactors = FALSE
)

as.key(nrc[nrc[["polarity"]] != 0, ])

bing <- syuzhet:::bing[!as.character(syuzhet:::bing[[1]]) \%in\%
    sentimentr::valence_shifters_table[["x"]], ]
sentiment(gsub("Sam-I-am", "Sam I am", sam_i_am), as.key(bing))
}
}
\keyword{hash}
\keyword{key}
\keyword{lookup}

