% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_key.R
\name{update_key}
\alias{is_key}
\alias{update_key}
\title{Convert data.frame to a Hash Key}
\usage{
update_key(key, drop = NULL, x = NULL,
  comparison = sentimentr::valence_shifters_table, sentiment = FALSE, ...)

is_key(key, sentiment = TRUE)
}
\arguments{
\item{key}{A \pkg{sentimentr} hash key.}

\item{drop}{A vector of terms to drop.}
}
\description{
\code{update_key} - Add/remove terms to a current key.

\code{is_key} - Logical check if an object is a key.
}

