% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine \pkg{sentimentr}'s Sentiment Data Sets}
\usage{
combine_data(
  data = c("course_evaluations", "hotel_reviews", "kaggle_movie_reviews",
    "kotzias_reviews_amazon_cells", "kotzias_reviews_imdb", "kotzias_reviews_yelp",
    "nyt_articles"),
  ...
)
}
\arguments{
\item{data}{A character vector of \pkg{sentimentr} data sets.}

\item{\ldots}{ignored.}
}
\value{
Returns an rbinded \pkg{data.table} of sentiment data with the source
added as column.
}
\description{
Combine trusted sentiment data sets from \pkg{sentimentr}.
}
\examples{
combine_data()
combine_data(c("kotzias_reviews_amazon_cells", "kotzias_reviews_imdb", 
    "kotzias_reviews_yelp"))
}
