% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{crowdflower_deflategate}
\alias{crowdflower_deflategate}
\title{Twitter Tweets About the Deflategate}
\format{
A data frame with 11,786 rows and 2 variables
}
\usage{
data(crowdflower_deflategate)
}
\description{
A dataset containing Twitter tweets about Tom Brady's deflated ball scandal, 
taken from Crowdflower.
}
\details{
\itemize{ 
  \item sentiment. A human scoring of the text.
  \item text. The sentences from the tweet.
}
}
\references{
Original URL: https://www.crowdflower.com/data-for-everyone
}
\keyword{datasets}
