% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{crowdflower_products}
\alias{crowdflower_products}
\title{Twitter Tweets About the Products}
\format{
A data frame with 3,548 rows and 2 variables
}
\usage{
data(crowdflower_products)
}
\description{
A dataset containing Twitter tweets about various products, taken from 
Crowdflower.
}
\details{
\itemize{ 
  \item sentiment. A human scoring of the text.
  \item text. The sentences from the tweet.
}
}
\references{
Cavender-Bares, K., (2013). Judge emotion about brands & products. \cr \cr
Original URL: https://www.crowdflower.com/data-for-everyone
}
\keyword{datasets}
