% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{crowdflower_self_driving_cars}
\alias{crowdflower_self_driving_cars}
\title{Twitter Tweets About Self Driving Cars}
\format{
A data frame with 6,943 rows and 2 variables
}
\usage{
data(crowdflower_self_driving_cars)
}
\description{
A dataset containing Twitter tweets about self driving cars, taken from
Crowdflower.
}
\details{
\itemize{
  \item sentiment. A human scoring of the text.
  \item text. The sentences from the tweet.
}
}
\references{
Original URL: https://www.crowdflower.com/data-for-everyone
}
\keyword{datasets}
