% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{crowdflower_weather}
\alias{crowdflower_weather}
\title{Twitter Tweets About the Weather}
\format{
A data frame with 763 rows and 2 variables
}
\usage{
data(crowdflower_weather)
}
\description{
A dataset containing Twitter tweets about the weather, taken from
Crowdflower.
}
\details{
\itemize{ 
  \item sentiment. A human scoring of the text.
  \item text. The sentences from the tweet. 
}
}
\references{
Original URL: https://www.crowdflower.com/data-for-everyone
}
\keyword{datasets}
