% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downweighted_zero_average.R
\name{downweighted_zero_average}
\alias{downweighted_zero_average}
\title{Downweighted Zeros Averaging}
\usage{
downweighted_zero_average(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{logical.  Should \code{NA} values should be stripped before the 
computation proceeds.}

\item{\ldots}{ignored.}
}
\value{
Returns a scalar summary of the zero downweighted average
}
\description{
Downweight the zeros in a vector for averaging.  This is sueful in the 
context of language where we don't want the neutral sentences to have such a 
strong influence on the general sentiment of the discourse with multiple 
sentences.  Essentially, this means neutral sentences are seen as having less 
emotional impact than a polarized sentence.
}
\examples{
x <- c(1, 2, 0, 0, 0, -1)
mean(x)
downweighted_zero_average(x)
downweighted_zero_average(c(NA, x))
mean(c(0, 0, 0, x))
downweighted_zero_average(c(0, 0, 0, x))
}

