% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{emoticons}
\alias{emoticons}
\title{Emoticons Data Set}
\format{A data frame with 75 rows and 2 variables}
\usage{
data(emoticons)
}
\description{
A dataset containing common emoticons (adapted from
\href{http://www.lingo2word.com/lists/emoticon_listH.html}{Popular Emoticon List}).
}
\details{
\itemize{
  \item x. The graphic representation of the emoticon
  \item y. The meaning of the emoticon
}
}
\references{
\url{http://www.lingo2word.com/lists/emoticon_listH.html}
}
\keyword{datasets}

