% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_rescale.R
\name{general_rescale}
\alias{general_rescale}
\title{Rescale a Numeric Vector}
\usage{
general_rescale(x, lower = -1, upper = 1, keep.zero = TRUE,
  sign = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{lower}{An upper limit to rescale to.}

\item{upper}{A lower limit to rescale to.}

\item{keep.zero}{logical.  If \code{TRUE} the zeros are kept as neutral.}

\item{sign}{logical.  If \code{TRUE} the vector will be scaled as (-1, 1, or 0)}

\item{\ldots}{ignored.}
}
\value{
Returns a rescaled vector of the same length as \code{x}.
}
\description{
Rescale a numeric vector with the option to make signed (-1, 1, or 0) and 
retain zero as neutral.
}
\examples{
general_rescale(c(NA, -3:3))
general_rescale(c(NA, -3:3), binary = TRUE)
general_rescale(c(NA, -3:3), keep.zero = FALSE)
general_rescale(c(NA, -3:3), keep.zero = FALSE, lower = 0, upper = 100)
}

