% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sentences.R
\name{get_sentences}
\alias{get_sentences}
\title{Get Sentences}
\usage{
get_sentences(x, ...)
}
\arguments{
\item{x}{A character vector, \code{sentiment}, or \code{sentiment_by} object.}

\item{\ldots}{Ignored.}
}
\value{
\code{subs} - Returns a list of sub-expressions.
}
\description{
Get sentences from a character vector, \code{sentiment}, or
\code{sentiment_by} object.
}
\examples{
(x <- paste0(
    "Mr. Brown comes! He says hello. i give him coffee.  i will ",
    "go at 5 p. m. eastern time.  Or somewhere in between!go there"
))
get_sentences(x)
}

