% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{get_sentences2}
\alias{get_sentences2}
\title{Get Sentences (Deprecated)}
\usage{
get_sentences2(x, ...)
}
\arguments{
\item{x}{A character vector, \code{sentiment}, or \code{sentiment_by} object.}

\item{\ldots}{Ignored.}
}
\description{
Deprecated, use \code{\link[sentimentr]{get_sentences}} instead.
}
