% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{grades}
\alias{grades}
\title{Grades Data Set}
\format{A data frame with 15 rows and 2 variables}
\usage{
data(grades)
}
\description{
A dataset containing common grades.
}
\details{
\itemize{
  \item x. The graphic representation of the grade
  \item y. The meaning of the grade
}
}
\keyword{datasets}

