% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{hotel_reviews}
\alias{hotel_reviews}
\title{Hotel Reviews}
\format{
A data frame with 5000 rows and 2 variables
}
\usage{
data(hotel_reviews)
}
\description{
A dataset containing a random sample (n = 5000 of 1,621,956) of Wang, Lu, & 
Zhai's (2011) hotel reviews data set scraped by the authors from 
Original URL: http://www.tripadvisor.com.
}
\details{
\itemize{ 
  \item sentiment. The overall rating for the experience
  \item text. The text review of the hotel
}
}
\references{
Wang, H., Lu, Y., and Zhai, C. (2011). Latent aspect rating 
analysis without aspect keyword supervision. In Proceedings of the 17th ACM 
SIGKDD Conference on Knowledge Discovery and Data Mining (KDD'2011), 618-626. \cr \cr
Original URL: 'http://sifaka.cs.uiuc.edu/~wang296/Data/index.html'
}
\keyword{datasets}
