% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{kaggle_movie_reviews}
\alias{kaggle_movie_reviews}
\title{Movie Reviews}
\format{
A data frame with 7,086 rows and 2 variables
}
\usage{
data(kaggle_movie_reviews)
}
\description{
A dataset containing sentiment scored movie reviews from a Kaggle competition
posted by University of Michigan SI650.  The data was originally collected 
from opinmind.com.
}
\details{
\itemize{ 
  \item sentiment. A numeric sentiment score
  \item text. The text from the review 
}
}
\references{
Original URL: https://www.kaggle.com/c/si650winter11/data
}
\keyword{datasets}
