% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion_by.R
\name{plot.emotion_by}
\alias{plot.emotion_by}
\title{Plots a emotion_by object}
\usage{
\method{plot}{emotion_by}(x, ordered = TRUE, ...)
}
\arguments{
\item{x}{The emotion_by object.}

\item{ordered}{logical.  If \code{TRUE} order the output grouping by emotion.}

\item{\ldots}{ignored}
}
\value{
Returns a \pkg{ggplot2} object.
}
\description{
Plots a emotion_by object.  Red centers are average emotion.  Alpha
jittered dots are raw sentence level emotion data.  Boxes are boxplots.
}
