% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity.R
\name{plot.profanity}
\alias{plot.profanity}
\title{Plots a profanity object}
\usage{
\method{plot}{profanity}(x, transformation.function = syuzhet::get_dct_transform, ...)
}
\arguments{
\item{x}{The profanity object.}

\item{transformation.function}{A transformation function to smooth the profanity
scores.}

\item{\ldots}{Other arguments passed to \code{\link[syuzhet]{get_transformed_values}}.}
}
\value{
Returns a \pkg{ggplot2} object.
}
\description{
Plots a profanity object.
}
\details{
Utilizes Matthew Jocker's \pkg{syuzhet} package to calculate smoothed
profanity across the duration of the text.
}
