% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_by.R
\name{plot.profanity_by}
\alias{plot.profanity_by}
\title{Plots a profanity_by object}
\usage{
\method{plot}{profanity_by}(x, ordered = TRUE, ...)
}
\arguments{
\item{x}{The profanity_by object.}

\item{ordered}{logical.  If \code{TRUE} order the output grouping by profanity.}

\item{\ldots}{ignored}
}
\value{
Returns a \pkg{ggplot2} object.
}
\description{
Plots a profanity_by object.  Red centers are average profanity.  Alpha
jittered dots are raw sentence level profanity data.  Boxes are boxplots.
}
