% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sentiment.R
\name{plot.sentiment}
\alias{plot.sentiment}
\title{Plots a sentiment object}
\usage{
\method{plot}{sentiment}(x, ...)
}
\arguments{
\item{x}{The sentiment object.}

\item{\ldots}{Other arguments passed to \code{\link[syuzhet]{get_transformed_values}}.}
}
\value{
Returns a \pkg{ggplot2} object.
}
\description{
Plots a sentiment object.
}
\details{
Utilizes Matthew Jocker's \pkg{syuzhet} package to calculate smoothed
sentiment across the duration of the text.
}

