% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_by.R
\name{plot.sentiment_by}
\alias{plot.sentiment_by}
\title{Plots a sentiment_by object}
\usage{
\method{plot}{sentiment_by}(x, ordered = TRUE, ...)
}
\arguments{
\item{x}{The sentiment_by object.}

\item{ordered}{logical.  If \code{TRUE} order the output grouping by sentiment.}

\item{\ldots}{ignored}
}
\value{
Returns a \pkg{ggplot2} object.
}
\description{
Plots a sentiment_by object.  Red centers are average sentiment.  Alpha
jittered dots are raw sentence level sentiment data.  Boxes are boxplots.
}
