% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{presidential_debates_2012}
\alias{presidential_debates_2012}
\title{2012 U.S. Presidential Debates}
\format{
A data frame with 2912 rows and 4 variables
}
\usage{
data(presidential_debates_2012)
}
\description{
A dataset containing a cleaned version of all three presidential debates for
the 2012 election.
}
\details{
\itemize{
  \item person. The speaker
  \item tot. Turn of talk
  \item dialogue. The words spoken
  \item time. Variable indicating which of the three debates the dialogue is from
}
}
\keyword{datasets}
