% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_emotion_terms.R
\name{print.extract_emotion_terms}
\alias{print.extract_emotion_terms}
\title{Prints an extract_emotion_terms Object}
\usage{
\method{print}{extract_emotion_terms}(x, ...)
}
\arguments{
\item{x}{An extract_emotion_terms object.}

\item{\ldots}{ignored}
}
\description{
Prints an extract_emotion_terms object
}
