% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_profanity_terms.R
\name{print.extract_profanity_terms}
\alias{print.extract_profanity_terms}
\title{Prints an extract_profanity_terms Object}
\usage{
\method{print}{extract_profanity_terms}(x, ...)
}
\arguments{
\item{x}{An extract_profanity_terms object.}

\item{\ldots}{ignored}
}
\description{
Prints an extract_profanity_terms object
}
