% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sentiment_terms.R
\name{print.extract_sentiment_terms}
\alias{print.extract_sentiment_terms}
\title{Prints an extract_sentiment_terms Object}
\usage{
\method{print}{extract_sentiment_terms}(x, ...)
}
\arguments{
\item{x}{An extract_sentiment_terms object.}

\item{\ldots}{ignored}
}
\description{
Prints an extract_sentiment_terms object
}
