% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sentiment.R
\name{print.validate_sentiment}
\alias{print.validate_sentiment}
\title{Prints a validate_sentiment Object}
\usage{
\method{print}{validate_sentiment}(x, ...)
}
\arguments{
\item{x}{A \code{validate_sentiment} Object}

\item{\ldots}{ignored.}
}
\description{
Prints a validate_sentiment object
}
