% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{ratings}
\alias{ratings}
\title{Ratings Data Set}
\format{A data frame with 35 rows and 2 variables}
\usage{
data(ratings)
}
\description{
A dataset containing common ratings.
}
\details{
\itemize{
  \item x. The graphic representation of the rating
  \item y. The meaning of the rating
}
}
\keyword{datasets}

