% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textclean_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{replace_grade}
\alias{reexports}
\alias{replace_rating}
\alias{reexports}
\alias{replace_emoticon}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{textclean}{\code{\link[textclean]{replace_grade}}, \code{\link[textclean]{replace_rating}}, \code{\link[textclean]{replace_emoticon}}}
}}

