% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_grade.R
\name{replace_grade}
\alias{replace_grade}
\title{Replace Grades With Words}
\usage{
replace_grade(text.var, grade_dt = sentimentr::grades, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grade_dt}{A \pkg{data.table} of gradess and corresponding word meanings.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector of strings with grades replaced with word
equivalents.
}
\description{
Replaces gradess with word equivalents.
}
\examples{
(text <- replace_grade(c(
    "I give an A+",
    "He deserves an F",
    "It's C+ work",
    "A poor example deserves a C!"
)))
sentiment(text)
}
\keyword{grade}

