% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{sam_i_am}
\alias{sam_i_am}
\title{Sam I Am Text}
\format{
A character vector with 169 elements
}
\usage{
data(sam_i_am)
}
\description{
A dataset containing a character vector of the text from Seuss's 'Sam I Am'.
}
\references{
Seuss, Dr. (1960). Green Eggs and Ham.
}
\keyword{datasets}
