% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{package}
\name{sentimentr}
\alias{package-sentiment}
\alias{sentimentr}
\alias{sentimentr-package}
\title{Calulate Text Polarity Sentiment}
\description{
Calulate text polarity sentiment at the sentence level and optionally
aggregate by rows or grouping variable(s).
}

