% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_key.R
\name{update_polarity_table}
\alias{update_polarity_table}
\title{Convert data.frame to a Hash Key}
\usage{
update_polarity_table(key, drop = NULL, x = NULL,
  comparison = sentimentr::valence_shifters_table, sentiment = FALSE, ...)
}
\description{
\code{update_polarity_table} - Wrapper for \code{update_key} specifically for
updating polarity tables.
}

