% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_key.R
\name{update_valence_shifter_table}
\alias{update_valence_shifter_table}
\title{Convert data.frame to a Hash Key}
\usage{
update_valence_shifter_table(key, drop = NULL, x = NULL,
  comparison = sentimentr::polarity_table, sentiment = FALSE, ...)
}
\description{
\code{update_valence_shifter_table} - Wrapper for \code{update_key} 
specifically for updating valence shifter tables.
}

