% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_key.R
\name{as_key}
\alias{as_key}
\alias{is_key}
\alias{update_key}
\title{Create/Manipulate Hash Keys}
\usage{
as_key(x, comparison = sentimentr::valence_shifters_table, ...)

update_key(key, drop = NULL, x = NULL,
  comparison = sentimentr::valence_shifters_table, ...)

is_key(key)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} with the first column containing
polarized words and the second containing polarity values.}

\item{comparison}{A \code{\link[base]{data.frame}} to compare to \code{x}.
If elements in \code{x}'s column 1 matches \code{comparison}'s column 1 the
accompanying row will be removed from \code{x}.  This is useful to ensure
\code{polarity_dt} words are not also found in \code{valence_shifters_dt} in
\code{\link[sentimentr]{sentiment}}.  Use \code{comparison = NULL} to skip
this comparison.}

\item{key}{A \pkg{sentimentr} hash key.}

\item{drop}{A vector of terms to drop.}

\item{\ldots}{ignored.}
}
\value{
Returns a \pkg{data.table} object that can be used as a hash key.
}
\description{
\code{as_key} - Create your own hash keys from a data frame for use in key
arguments such as \code{polarity_dt} in the \code{sentiment} function.

\code{update_key} - Add/remove terms to a current key.

\code{is_key} -
}
\examples{
key <- data.frame(
    words = sample(LETTERS),
    polarity = rnorm(26),
    stringsAsFactors = FALSE
)

(mykey <- as_key(key))

## Looking up values
mykey[c("A", "K")][[2]]

## Drop terms from key
update_key(mykey, drop = c("F", "H"))

## Add terms to key
update_key(mykey, x = data.frame(x = c("Dog", "Cat"), y = c(1, -1)))

## Add terms & drop to/from a key
update_key(mykey, drop = c("F", "H"), x = data.frame(x = c("Dog", "Cat"), y = c(1, -1)))

## Checking if you have a key
is_key(mykey)
is_key(key)
is_key(mtcars)
is_key(update_key(mykey, drop = c("F", "H")))

## Using syuzhet's sentiment lexicons
\dontrun{
library(syuzhet)
as_key(syuzhet:::bing)
as_key(syuzhet:::afinn)
nrc <- data.frame(
    words = rownames(syuzhet:::nrc),
    polarity = syuzhet:::nrc[, "positive"] - syuzhet:::nrc[, "negative"],
    stringsAsFactors = FALSE
)

as_key(nrc[nrc[["polarity"]] != 0, ])

sentiment(gsub("Sam-I-am", "Sam I am", sam_i_am), as_key(syuzhet:::bing))
}
}
\keyword{hash}
\keyword{key}
\keyword{lookup}

