% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{data}
\name{polarity_table}
\alias{polarity_table}
\title{Polarity Lookup Key}
\format{A data frame with 6821 rows and 2 variables}
\usage{
data(polarity_table)
}
\description{
A \pkg{data.table} dataset containing an augmented version of Hu & Liu's (2004)
positive/negative word list as sentiment lookup values.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment values (+1, -1)
}
}
\references{
Hu, M., & Liu, B. (2004). Mining opinion features in customer
reviews. National Conference on Artificial Intelligence.

\file{https://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\keyword{datasets}

