% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textclean_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{replace_grade}
\alias{replace_rating}
\alias{replace_emoticon}
\alias{replace_emoji}
\alias{replace_emoji_identifier}
\alias{replace_internet_slang}
\alias{replace_word_elongation}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{textclean}{\code{\link[textclean]{replace_emoji}}, \code{\link[textclean:replace_emoji]{replace_emoji_identifier}}, \code{\link[textclean]{replace_emoticon}}, \code{\link[textclean]{replace_grade}}, \code{\link[textclean]{replace_internet_slang}}, \code{\link[textclean]{replace_rating}}, \code{\link[textclean]{replace_word_elongation}}}
}}

