% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentr-package.R
\docType{package}
\name{sentimentr}
\alias{sentimentr}
\alias{package-sentiment}
\title{Calculate Text Polarity Sentiment}
\description{
Calculate text polarity sentiment in the English language at the sentence 
level and optionally aggregate by rows or grouping variable(s).
}
