% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sentiment.R
\name{validate_sentiment}
\alias{validate_sentiment}
\title{Validate Sentiment Score Sign Against Known Results}
\usage{
validate_sentiment(predicted, actual, ...)
}
\arguments{
\item{predicted}{A numeric vector of predicted sentiment scores or a 
\pkg{sentimentr} object that returns sentiment scores.}

\item{actual}{A numeric vector of known sentiment ratings.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{\link[base]{data.frame}} with a macroaveraged and 
microaveraged model validation scores.  Additionally, the
 \code{\link[base]{data.frame}} has the following attributes:
\item{confusion_matrix}{A confusion matrix of all classes}
\item{class_confusion_matrices}{A \code{\link[base]{list}} of class level (class vs. all) confusion matrices}
\item{macro_stats}{A \code{\link[base]{data.frame}} of the macroaverged class level stats before averaging}
}
\description{
Provides a miulticlass macroaverage/microaverage of precision, recall, 
accuracy, and F-score for the sign of the predicted sentiment against known 
sentiment scores. Macroaveraging allows every class to have an equal say.  
Microaveraging gives larger say to larger classes.
}
\examples{
actual <- c(1, 1, 1, 1, -1, -1, -1, -1, -1, -1, -1, 1,-1)
predicted <- c(1, 0, 1, -1, 1, 0, -1, -1, -1, -1, 0, 1,-1)
validate_sentiment(actual, predicted)

scores <- cannon_reviews$opinion.score
mod <- sentiment_by(get_sentences(cannon_reviews$review))

validate_sentiment(mod$ave_sentiment, scores)
validate_sentiment(mod, scores)

x <- validate_sentiment(mod, scores)
attributes(x)$confusion_matrix
attributes(x)$class_confusion_matrices
attributes(x)$macro_stats
}
\references{
\url{https://www.youtube.com/watch?v=OwwdYHWRB5E&index=31&list=PL6397E4B26D00A269}
}
